+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/LanguageSPITest.java
@@ -784,7 +784,6 @@ public class LanguageSPITest {
 
         List<Env> createContextCalled = new ArrayList<>();
         List<Source> parseCalled = new ArrayList<>();
-        List<Integer> initializeMultiContextCalled = new ArrayList<>(); // executionIndex
         List<Integer> initializeMultipleContextsCalled = new ArrayList<>(); // executionIndex
         boolean contextCachingEnabled = false;
         int executionIndex;
@@ -834,13 +833,6 @@ public class LanguageSPITest {
             initializeMultipleContextsCalled.add(executionIndex);
         }
 
-        @Override
-        protected boolean initializeMultiContext() {
-            executionIndex++;
-            initializeMultiContextCalled.add(executionIndex);
-            return contextCachingEnabled;
-        }
-
         static MultiContextLanguage getInstance(Class<? extends MultiContextLanguage> lang, Context context) {
             context.enter();
             try {
@@ -868,7 +860,7 @@ public class LanguageSPITest {
         MultiContextLanguage lang = MultiContextLanguage.getInstance(MultiContextLanguage.class, context);
 
         assertEquals(1, lang.createContextCalled.size());
-        assertTrue(lang.initializeMultiContextCalled.isEmpty());
+        assertTrue(lang.initializeMultipleContextsCalled.isEmpty());
         assertTrue(lang.parseCalled.isEmpty());
 
         context.eval(source1);
@@ -877,12 +869,12 @@ public class LanguageSPITest {
 
         context.eval(source1); // cached parse
         assertEquals(1, lang.parseCalled.size());
-        assertTrue(lang.initializeMultiContextCalled.isEmpty());
+        assertTrue(lang.initializeMultipleContextsCalled.isEmpty());
 
         context.eval(source2); // cached parse
         assertEquals(2, lang.parseCalled.size());
         assertEquals(source2.getCharacters(), lang.parseCalled.get(1).getCharacters());
-        assertTrue(lang.initializeMultiContextCalled.isEmpty());
+        assertTrue(lang.initializeMultipleContextsCalled.isEmpty());
         context.close();
     }
 
@@ -903,7 +895,7 @@ public class LanguageSPITest {
         MultiContextLanguage outerLang = MultiContextLanguage.getInstance(MultiContextLanguage.class, context);
 
         assertEquals(1, outerLang.createContextCalled.size());
-        assertTrue(outerLang.initializeMultiContextCalled.isEmpty());
+        assertTrue(outerLang.initializeMultipleContextsCalled.isEmpty());
         assertTrue(outerLang.parseCalled.isEmpty());
         Env env = outerLang.createContextCalled.get(0);
 
@@ -942,7 +934,7 @@ public class LanguageSPITest {
 
         context.leave();
         context.close();
-        assertEquals(0, outerLang.initializeMultiContextCalled.size());
+        assertEquals(0, outerLang.initializeMultipleContextsCalled.size());
         assertEquals(1, outerLang.createContextCalled.size());
     }
 
@@ -961,7 +953,7 @@ public class LanguageSPITest {
         context.initialize(OneContextLanguage.ID);
         MultiContextLanguage lang = OneContextLanguage.getInstance(OneContextLanguage.class, context);
         assertEquals(1, lang.createContextCalled.size());
-        assertTrue(lang.initializeMultiContextCalled.isEmpty());
+        assertTrue(lang.initializeMultipleContextsCalled.isEmpty());
         assertTrue(lang.parseCalled.isEmpty());
         Env env = lang.createContextCalled.get(0);
 
@@ -978,7 +970,7 @@ public class LanguageSPITest {
         Env innerEnv = OneContextLanguage.getContext().env;
         innerEnv.parsePublic(truffleSource1);
         assertEquals(1, innerLang.parseCalled.size());
-        assertEquals(0, innerLang.initializeMultiContextCalled.size());
+        assertEquals(0, innerLang.initializeMultipleContextsCalled.size());
         assertEquals(1, innerLang.createContextCalled.size());
 
         innerEnv.parsePublic(truffleSource1);
@@ -998,7 +990,7 @@ public class LanguageSPITest {
         context.close();
 
         assertEquals(1, lang.createContextCalled.size());
-        assertEquals(0, lang.initializeMultiContextCalled.size());
+        assertEquals(0, lang.initializeMultipleContextsCalled.size());
     }
 
     private static Source getTruffleSource(org.graalvm.polyglot.Source source) throws NoSuchFieldException, IllegalAccessException {
@@ -1017,8 +1009,10 @@ public class LanguageSPITest {
         context2.initialize(OneContextLanguage.ID);
         MultiContextLanguage lang1 = OneContextLanguage.getInstance(OneContextLanguage.class, context1);
         MultiContextLanguage lang2 = OneContextLanguage.getInstance(OneContextLanguage.class, context2);
-        assertEquals(0, lang1.initializeMultiContextCalled.size());
-        assertEquals(0, lang2.initializeMultiContextCalled.size());
+        assertEquals(0, lang1.initializeMultipleContextsCalled.size());
+        assertEquals(0, lang2.initializeMultipleContextsCalled.size());
+        context1.close();
+        context2.close();
     }
 
     @Test
@@ -1027,10 +1021,8 @@ public class LanguageSPITest {
         Context context = Context.newBuilder().engine(engine).build();
         context.initialize(MultiContextLanguage.ID);
         MultiContextLanguage lang = MultiContextLanguage.getInstance(MultiContextLanguage.class, context);
-        assertEquals(1, lang.initializeMultiContextCalled.size());
         assertEquals(1, lang.initializeMultipleContextsCalled.size());
         assertEquals(1, (int) lang.initializeMultipleContextsCalled.get(0));
-        assertEquals(2, (int) lang.initializeMultiContextCalled.get(0));
         assertEquals(1, lang.createContextCalled.size());
         context.close();
         engine.close();
@@ -1049,10 +1041,8 @@ public class LanguageSPITest {
         MultiContextLanguage lang1 = MultiContextLanguage.getInstance(MultiContextLanguage.class, context1);
 
         assertTrue(lang1.parseCalled.isEmpty());
-        assertEquals(1, lang1.initializeMultiContextCalled.size());
         assertEquals(1, lang1.initializeMultipleContextsCalled.size());
         assertEquals(1, (int) lang1.initializeMultipleContextsCalled.get(0));
-        assertEquals(2, (int) lang1.initializeMultiContextCalled.get(0));
         assertEquals(1, lang1.createContextCalled.size());
 
         context1.eval(source1);
@@ -1092,10 +1082,8 @@ public class LanguageSPITest {
         assertEquals(1, lang2.createContextCalled.size());
         engine.close();
 
-        assertEquals(1, lang1.initializeMultiContextCalled.size());
         assertEquals(1, lang1.initializeMultipleContextsCalled.size());
         assertEquals(1, (int) lang1.initializeMultipleContextsCalled.get(0));
-        assertEquals(2, (int) lang1.initializeMultiContextCalled.get(0));
         assertEquals(1, lang1.createContextCalled.size());
         assertEquals(1, lang2.createContextCalled.size());
     }
@@ -1113,10 +1101,8 @@ public class LanguageSPITest {
         MultiContextLanguage lang = MultiContextLanguage.getInstance(MultiContextLanguage.class, context1);
 
         assertTrue(lang.parseCalled.isEmpty());
-        assertEquals(1, lang.initializeMultiContextCalled.size());
         assertEquals(1, lang.initializeMultipleContextsCalled.size());
         assertEquals(1, (int) lang.initializeMultipleContextsCalled.get(0));
-        assertEquals(2, (int) lang.initializeMultiContextCalled.get(0));
         assertEquals(1, lang.createContextCalled.size());
 
         context1.eval(source1);
@@ -1147,7 +1133,6 @@ public class LanguageSPITest {
         engine.close();
 
         assertEquals(2, lang.createContextCalled.size());
-        assertEquals(1, lang.initializeMultiContextCalled.size());
         assertEquals(1, lang.initializeMultipleContextsCalled.size());
     }
 
@@ -1223,34 +1208,6 @@ public class LanguageSPITest {
         c.initialize(ProxyLanguage.ID);
         testFails(() -> c.eval(ProxyLanguage.ID, "t0"));
         testFails(() -> c.eval(ProxyLanguage.ID, "t1"));
-        c.close();
-    }
-
-    @Test
-    public void testErrorInFindMetaObject() {
-        final TruffleObject testObject = new TruffleObject() {
-        };
-        ProxyLanguage.setDelegate(new ProxyLanguage() {
-
-            @Override
-            protected Object findMetaObject(LanguageContext context, Object value) {
-                throw new RuntimeException();
-            }
-
-            @Override
-            protected CallTarget parse(com.oracle.truffle.api.TruffleLanguage.ParsingRequest request) throws Exception {
-                return RootNode.createConstantNode(testObject).getCallTarget();
-            }
-
-            @Override
-            protected boolean isObjectOfLanguage(Object object) {
-                return object == testObject;
-            }
-        });
-        Context c = Context.create();
-        Value v = c.eval(ProxyLanguage.ID, "");
-        testFails(() -> v.getMetaObject());
-        testFails(() -> v.getMetaObject());
         c.close();
     }
 
@@ -1943,78 +1900,6 @@ public class LanguageSPITest {
         c.close();
     }
 
-    @Test
-    public void testFindSourceLocation() {
-        ProxyLanguage.setDelegate(new ProxyLanguage() {
-            @Override
-            protected CallTarget parse(TruffleLanguage.ParsingRequest request) throws Exception {
-                return RootNode.createConstantNode(new SourceHolder(request.getSource())).getCallTarget();
-            }
-
-            @Override
-            protected SourceSection findSourceLocation(ProxyLanguage.LanguageContext context, Object value) {
-                if (value instanceof SourceHolder) {
-                    final Source src = ((SourceHolder) value).source;
-                    return src.createSection(0, src.getLength());
-                }
-                return super.findSourceLocation(context, value);
-            }
-
-            @Override
-            protected boolean isObjectOfLanguage(Object object) {
-                return object instanceof SourceHolder;
-            }
-        });
-
-        try (Context context = Context.create(ProxyLanguage.ID)) {
-            String text = "01234567";
-            Value res = context.eval(ProxyLanguage.ID, text);
-            assertNotNull(res);
-            org.graalvm.polyglot.SourceSection sourceSection = res.getSourceLocation();
-            assertNotNull(sourceSection);
-            assertTrue(text.contentEquals(sourceSection.getCharacters()));
-            res = context.asValue(new SourceHolder(Source.newBuilder(ProxyLanguage.ID, text, null).build()));
-            sourceSection = res.getSourceLocation();
-            assertNotNull(sourceSection);
-            assertTrue(text.contentEquals(sourceSection.getCharacters()));
-        }
-    }
-
-    @Test
-    public void testToString() {
-        ProxyLanguage.setDelegate(new ProxyLanguage() {
-            @Override
-            protected CallTarget parse(TruffleLanguage.ParsingRequest request) throws Exception {
-                return RootNode.createConstantNode(new SourceHolder(request.getSource())).getCallTarget();
-            }
-
-            @Override
-            protected String toString(ProxyLanguage.LanguageContext context, Object value) {
-                if (value instanceof SourceHolder) {
-                    final Source src = ((SourceHolder) value).source;
-                    return src.getCharacters().toString();
-                }
-                return super.toString(context, value);
-            }
-
-            @Override
-            protected boolean isObjectOfLanguage(Object object) {
-                return object instanceof SourceHolder;
-            }
-        });
-
-        try (Context context = Context.create(ProxyLanguage.ID)) {
-            String text = "01234567";
-            Value res = context.eval(ProxyLanguage.ID, text);
-            assertNotNull(res);
-            String toString = res.toString();
-            assertEquals(text, toString);
-            res = context.asValue(new SourceHolder(Source.newBuilder(ProxyLanguage.ID, text, null).build()));
-            toString = res.toString();
-            assertEquals(text, toString);
-        }
-    }
-
     private static boolean lookupLanguage(Class<?> serviceClass) {
         Env env = com.oracle.truffle.api.test.polyglot.ProxyLanguage.LanguageContext.get(null).env;
         LanguageInfo languageInfo = env.getInternalLanguages().get(SERVICE_LANGUAGE);
@@ -2148,16 +2033,6 @@ public class LanguageSPITest {
         final Engine engine = context.getEngine();
         assertEquals(engine.getVersion(), engine.getLanguages().get(INHERITED_VERSION).getVersion());
         context.close();
-    }
-
-    private static class SourceHolder implements TruffleObject {
-        final Source source;
-
-        SourceHolder(final Source source) {
-            Objects.requireNonNull(source, "The source must be non null.");
-            this.source = source;
-        }
-
     }
 
     @SuppressWarnings("serial")