+++ truffle/CHANGELOG.md
@@ -30,6 +30,7 @@ No active inner context is allowed after `TruffleLanguage.finalizeContext(Object
 * Removed deprecated object DSL processor that was deprecated for several releases. 
 * Removed deprecated encapsulating node accessor methods in `NodeUtil`.
 * Removed deprecated method `LoopNode.executeLoop`.
+* Removed deprecated many methods like getMetaObject and toString in TruffleLanguage deprecated in 20.1.
 
 * Added new APIs to `com.oracle.truffle.api.frame.Frame` and `com.oracle.truffle.api.frame.FrameDescriptor`:
  * Added a new "namespace" of index-based slots in `Frame` that is defined during construction of the frame descriptor and cannot be changed afterwards, and that is accessed using `int` indexes instead of `FrameSlot`s.